+++ src/Pages/DetailPage/DetailBody.tsx
