+++ src/Pages/DetailPage/DetailBottom.tsx
