+++ src/Pages/DetailPage/DetailHeader.tsx
