+++ src/Pages/DetailPage/DetailPartySection.tsx
