+++ src/Pages/DetailPage/style.ts
