+++ src/Pages/MainPage/Components/BottomSheet/BottomSheetHandle.tsx
