+++ src/main.tsx
@@ -2,8 +2,8 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import App from './App';
-import MainPage from './MainPage/MainPage';
-import DetailPage from './MainPage/DetailPage';
+import MainPage from './Pages/MainPage/MainPage';
+import DetailPage from './Pages/DetailPage';
 
 const router = createBrowserRouter([
   {
